Add LayoutController.php and Layout routes
Added the Layout Controller file, as well as some basic routes. The Layout Test file needs to be updated with its own logic, as the current logic mirroring how table view and action bars are handled will not work. This should be a simpler fix as the logic is less intensive.