Update User Table Seeder data
Added users.json file to contain test user data - in future it will contain out-of-the-box user data.

Updated layouts.json Owner field properties to be processable by the frontend.

Updated the UsersTableSeeder.php file to add the test user data. This will need to be further modified if multiple custom users are to be added.